snap noise off by default; another improvement to the OpenDialog
Turn the snap noise off by default. Can still enable in settings menu. Also, another improvement to the Open File dialog. Tapping out was so smooth when we just kept the current file open. Alas, this caused occasional problems when the file did not close fast enough to delete it or open another file. Now we will just hold onto the name of the file that was open and reopen it upon tap out.