Compass Calibration
Moved compass calibration to the settings menu (is now just one dialog for all connected robots). Added callback to update the calibration status and remove the video when calibration is complete. Adjusted the video element for better positioning and to work with Android. Added functions to remove videos as necessary.