Korean, language misc., buzzer bug,
Set the buzzer block to only send a command if the duration is > 0. This makes up for a bug in the micro:bit where the first buzzer command, if duration is 0, sets the buzzer to go forever. Added Korean translations plus some additional language keys requsted by the Korean translator. Changed 'Enter text' in the NumOrStringSlot popup to 'Edit text'. Added a debug check to make sure that all languages have the same keys as English (results displayed in browser console). Moved each language to its own file and moved all language files to a new folder. Removed 'files will remain' text, and text traslation if no project is open (should not happen). Reenabled the language option in the long press settings menu for debuging.