Spanish, when block, firmware error and more
Added the Spanish translation as well as updates to Catalan, Dutch, and French. Modified the behavior of the when <bool> block to match that in Snap!. Removed the incompatible firmware error in favor of the connection failed error. Updated the connection failed error to remove the failed device from the list automatically (before, you would have to manually disconnect before you could try again).