Sound block limits, misc.
Added limits to the number slots on two of the sound blocks. Changed the device label text to display correctly in RTL. Changed some Arabic text to display better in iOS. Fixed the Hebrew tri led block text so that it also displays correctly. Added the filename to the delete popup box so that it is clear what you are deleting.