order of operations, misc.
Corrected the order of operations in the subtraction and division blocks so that they read from left to right even in right to left languages. Removed the blank labels that were accidentially being added by the block parser. Added some explaination of the strange change I had to make to categories.