BlockButton and SliderWidget
Added a BlockPart that is a button - when pressed it will bring up an InputSystem to change the value on the button. This BlockPart is for FinchBlox and is not a Slot. No other blocks can be dragged into it. Also added a SliderWidget which can be added to any InputPad.