Finch motion blocks, Level Dialog, Finchblox finch button and dialog, misc.
If it's FinchBlox level 1 or 2, create a new flag block if the flag block is deleted. Updates to finch motion blocks. Beginnings of adding a shadow stack instead of a highlight when dragging in Finchblox. Added Avenir Heavy as the main font for FinchBlox. Removed the timer for the battery checker and added checking in appropriate locations. Changed FinchBlox level selection from a drop down menu to a dialog box. Added language files for Finnish and Swedish (still need Swedish translation). Updated look of Discover Dialog for FinchBlox. Improved function of finch button for FinchBlox. Fixed bugs in resizing and zooming.