FinchBlox V0.2 build 6. More FinchBlox Level 3
FinchBlox: Added the ability to make the top half of a block a different color. Adjusted the shape and stretching of the loop blocks. Made loop blocks appear smaller in the block palette so that they fit in the space. New icons for loop blocks and fixed the buttons on the loop and wait blocks. First draft of sensor blocks with darkness threshold at 5 and distance at 30 cm. Adjustments to zooming to accomidate different screen sizes. Disabled small mode since that has not been set up for FinchBlox and just looked strange.