Misc.
In FinchBlox, switch the ok and cancel buttons in the file name and delete popups. Add some time to the reset encoders block so that the finch has a chance to reset. Update limits on the birdblox finch move and turn blocks. Make sure the move and turn blocks don't move the finch when the distance is zero. Finch speed scaling update.