Add micro:bit V2 support
Add a new block for sound/temperature sensor in the micro:bit V2. This block should be grey if the device connected is not a V2, and should return a 'V2 micro:bit required' error. Add logo option to micro:bit button block which will return the same error if the the device is not a V2. Fix minor bug with the battery monitor title bar button.