Add recenter button and misc.
Add a button to FinchBlox to allow the user to recenter the view on all stacks with hat blocks. Outline the overflow arrows in FinchBlox. Fix new editable text bug.

Add recenter button and misc.

Add a button to FinchBlox to allow the user to recenter the view on all stacks with hat blocks. Outline the overflow arrows in FinchBlox. Fix new editable text bug.