Update hatchling
Update stop block for use with hatchling. Change the port button behavior so that it is only clickable if there are options, and it updates automatically on the canvas and in the block palette. Add the wait until light block for hatchling. Make block buttons visible in palette for hatchling.